% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI510.R
\name{df_scale}
\alias{df_scale}
\title{Returns a tibble x in which the numeric variables have been scaled with scale}
\usage{
df_scale(x, center = T, scale = T)
}
\arguments{
\item{x}{A tibble to scale}

\item{center}{Logical. Should the variables be centered? (default = TRUE)}

\item{scale}{Logical. Should the variables be scaled to have unit variance? (default = TRUE)}
}
\value{
A tibble with the numeric variables centered and/or scaled
}
\description{
Returns a tibble x in which the numeric variables have been scaled with scale
}
\examples{
df_scale(mtcars)

}
