% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI510.R
\name{is_max}
\alias{is_max}
\title{Check if values in a vector are equal to the max value}
\usage{
is_max(x, na.rm = T)
}
\arguments{
\item{x}{An atomic vector to check}

\item{na.rm}{Logical. Should missing values be removed? (default = TRUE)}
}
\value{
A logical vector with \code{TRUE} values where \code{x} equals its maximum value
}
\description{
This function accepts an atomic vector \code{x} and returns a logical vector with \code{TRUE} values where \code{x} equals its max value.
}
\examples{
is_max(c(3, 5, 1, 3, 2))
is_max(c(3, 5, NA, 3, 2))

}
