% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI510.R
\name{rando}
\alias{rando}
\title{Randomly sample elements or rows from a vector or data frame}
\usage{
rando(x, n = 1, replace = T)
}
\arguments{
\item{x}{An atomic vector, data frame, or tibble}

\item{n}{n samples will be sampled from X}

\item{replace}{default TRUE, this allows sampling with replacement}
}
\value{
vector or data frame
}
\description{
Randomly sample elements or rows from a vector or data frame
}
