% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMI510.R
\name{recall}
\alias{recall}
\title{This function calculates the recall of a binary classification model based on the predicted labels \code{pred} and the true labels \code{truth}.}
\usage{
recall(pred, truth)
}
\arguments{
\item{pred}{A binary vector of predicted labels}

\item{truth}{A binary vector of true labels}
}
\value{
The recall (sensitivity) of the binary classification model
}
\description{
This function calculates the recall of a binary classification model based on the predicted labels \code{pred} and the true labels \code{truth}.
}
\examples{
recall(c(1,1,0,1), c(1,0,0,1))

}
